now subscribes immediatley
